13
In this submission, I have resolved many errors.
After the two codes are merged, although the event addition is displayed successfully, it cannot be added to the database.
2. Clicking the client registration button cannot complete the registration.
When new events are added to the management end, the amount raised will always be 0 regardless of what is filled in, and the status will remain active regardless of whether it is checked or not.
4. After I changed the status through the edit button, an error was reported when I clicked the view registrations and Edit buttons.